Move button props to external composable to be able to build
